requirements.txt
# Computer Vision and ML
ultralytics
opencv-python
numpy
torch
torchvision
Pillow

# API and Web
openai
requests
flask

# Utilities
pyyaml
pytz
